PAR-43: removes unneccessary pipes and encapsulates Coordinates type
