+++ src/app/classes/diagram/coordinates.ts
@@ -1,0 +1,11 @@
+export type CoordinatesInfo = {
+    transitionName: string;
+    transitionType: string;
+    coordinates: Coordinates;
+    globalOffset: Coordinates;
+};
+
+export type Coordinates = {
+    x: number;
+    y: number;
+};
